fix(core): preserve wire timestamp + monotonic event id so bursts don't collapse
When several LXMF messages arrived in the same wall-clock second, only one was
shown. Two independent causes:

1. Wire timestamp discarded (primary). decode_lxmf_payload skipped msgpack
   element 0 and lxmf_event_from_bytes stamped SystemTime::now().as_secs(), so
   every message in a second got an identical second-resolution timestamp ->
   unstable order, conversation-preview collapse, and (source, timestamp) dedup
   merged them. Now the real LXMF wire timestamp is read (float64/float32/int ->
   f64, sub-second preserved) and carried through the event, persistence, and
   store. MessageReceived.timestamp is f64 seconds; the messages.timestamp
   column is REAL (SQLite affinity tolerates existing integer rows).

2. No per-event identity (buffer race). JS consumers tracked "what's new" by
   comparing the head event by reference against a 200-cap buffer; a burst that
   evicted the tracked head dropped or duplicated events. Every event now
   carries a strictly-increasing, unique id (AtomicU64, assigned once at drain/
   serialize time) so consumers can track lastSeenId. Package raises the buffer
   cap 200 -> 1000 and documents the id. Example app keys conversation bubbles
   by event id (was source-timestamp) and dedups live-vs-SQL by content.

Tests: +3 (burst keeps 3 distinct sub-second timestamps; float64/integer
timestamp values preserved); updated store/event tests for f64. 182 pass,
clippy clean.

+++ rust-core/src/store.rs
@@ -23,7 +23,7 @@ impl MessageStore {
                 image_data   BLOB,
                 files_json   TEXT,
                 outbound     INTEGER NOT NULL DEFAULT 0,
-                timestamp    INTEGER NOT NULL DEFAULT (strftime('%s','now')),
+                timestamp    REAL    NOT NULL DEFAULT (strftime('%s','now')),
                 acked        INTEGER NOT NULL DEFAULT 0
             );
             CREATE TABLE IF NOT EXISTS solana_txs (
@@ -70,7 +70,7 @@ impl MessageStore {
         body: &[u8],
         image: Option<(&str, &[u8])>,
         files: &[(String, Vec<u8>)],
-        timestamp: u64,
+        timestamp: f64,
     ) -> Result<i64, rusqlite::Error> {
         let image_mime: Option<&str> = image.map(|(m, _)| m);
         let image_data: Option<&[u8]> = image.map(|(_, d)| d);
@@ -99,7 +99,7 @@ impl MessageStore {
                 image_mime,
                 image_data,
                 files_json,
-                timestamp as i64,
+                timestamp,
             ],
         )?;
         Ok(conn.last_insert_rowid())
@@ -155,7 +155,7 @@ impl MessageStore {
             let image_data: Option<Vec<u8>> = row.get(6)?;
             let files_json: Option<String> = row.get(7)?;
             let outbound: i32              = row.get(8)?;
-            let timestamp: i64             = row.get(9)?;
+            let timestamp: f64             = row.get(9)?;
             let acked: i32                 = row.get(10)?;
             Ok((id, source_hex, dest_hex, title_bytes, body_bytes, image_mime, image_data, files_json, outbound, timestamp, acked))
         })?.filter_map(|r| r.ok()).map(|(id, source, dest, title_b, body_b, img_mime, img_data, files_j, outbound, ts, acked)| {
